+++ internal/domain/payment/dto/payment.go
@@ -6,17 +6,17 @@ import (
 )
 
 type CreatePaymentItem struct {
-	Name     string `json:"name" validate:"required"`
-	Price    int64  `json:"price" validate:"required,gte=0"`
-	Quantity int32  `json:"quantity" validate:"required,gte=0"`
-	Discount int32  `json:"discount" validate:"required"`
+	Name     string  `json:"name" validate:"required"`
+	Price    float32 `json:"price" validate:"required,gte=0"`
+	Quantity int32   `json:"quantity" validate:"required,gte=0"`
+	Discount int32   `json:"discount" validate:"required"`
 }
 
 type CreatePayment struct {
 	UserId    uuid.UUID `json:"userId" validate:"required,uuid4"`
 	OrderId   string    `json:"orderId" validate:"required"`
 	OrderInfo string    `json:"orderInfo" validate:"required"`
-	Amount    int64     `json:"amount" validate:"required"`
+	Amount    float32   `json:"amount" validate:"required"`
 
 	Items []CreatePaymentItem `json:"items" validate:"required,dive"`
 }
@@ -30,6 +30,6 @@ type UpdatePayment struct {
 	ID        int64                   `json:"id" validate:"required"`
 	OrderId   *string                 `json:"orderId" validate:"omitempty"`
 	OrderInfo *string                 `json:"orderInfo" validate:"omitempty"`
-	Amount    *int64                  `json:"amount" validate:"omitempty,gte=0"`
+	Amount    *float32                `json:"amount" validate:"omitempty,gte=0"`
 	Status    *database.PAYMENTSTATUS `json:"status" validate:"omitempty"`
 }
